[BE] refactor: mentoring dto 리팩토링
