--- server/src/domains/mentoring/dto/MentoringRequestListDto.ts
@@ -1,22 +1,0 @@
-import { IsDateString, IsNumber, IsOptional, IsString } from 'class-validator';
-
-export class MentoringRequestListDto {
-  @IsNumber()
-  id: number;
-  @IsNumber()
-  groupId: number;
-  @IsString()
-  groupName: string;
-  @IsString()
-  categoryImage: string;
-  @IsString()
-  ownerName: string;
-  @IsDateString()
-  createdAt: string;
-}
-
-export class MentoringRequestParam {
-  @IsOptional()
-  @IsNumber()
-  mid: number;
-}
