+++ server/src/domains/mentoring/dto/param/MentoringRequestParam.ts
@@ -1,0 +1,7 @@
+import { IsNumber, IsOptional } from 'class-validator';
+
+export class MentoringRequestParam {
+  @IsOptional()
+  @IsNumber()
+  mid: number;
+}
