+++ server/src/domains/mentoring/dto/request/AcceptRequestDto.ts
@@ -1,4 +1,4 @@
-import { IsNumber } from "class-validator";
+import { IsNumber } from 'class-validator';
 
 export class AcceptRequestDto {
   @IsNumber()
