+++ server/src/domains/mentoring/dto/response/MentorInfoResponse.ts
@@ -1,0 +1,6 @@
+import { IsNumber } from 'class-validator';
+
+export class MentorInfoResponse {
+  @IsNumber()
+  mentorId: number;
+}
